fix(docsync): back the sync attempt token with a reducer guard
Follow-up review of 54f4791.

- The reducer refuses terminal network actions on a settled query again.
  The attempt token in `handleSync` is the real fix for superseded syncs,
  but it is enforced by a check placed after every `await`, and mutation
  testing showed only one of those ten checks is covered by a test. Tracing
  each path shows the state-mutating steps are protected independently by
  `finishSyncAttempt`, the transaction's `isCurrent()` and
  `applyServerOperations`, so the remaining nine guard against spurious
  requests, events and broadcasts rather than corruption. Restoring
  `isSettled` puts back the one guard that catches a forgotten check and
  that the reducer matrix verifies.
- `networkQueryError` only accepts `"fetching"` as an override. It is the
  single value ever passed, and a wider type would let a caller force
  `idle` onto a paused query around `terminalFetchStatus`.
- The subdocs example gained the same inline error banners as the editor
  example, and the same silent-failure hole: the panels render alongside
  an error, and nothing asserted its absence. The banners now carry
  `data-testid="query-error"` and the subdocs UI helper asserts none is
  present, mirroring the editor helper.
- `pauseQueries` moves to `utils/`; a connection handler was also serving
  as a utility module for the client class.
