+++ packages/docsync/src/client/handlers/connection/disconnect.ts
@@ -1,35 +1,7 @@
 import type { DocSyncClient } from "../../index.js";
 import { DocSyncError } from "../../utils/DocSyncError.js";
-import {
-  dispatchDocQueryConnectionError,
-  dispatchDocQueryDisconnected,
-} from "../../utils/dispatchDocQueryAction.js";
+import { pauseQueries } from "../../utils/pauseQueries.js";
 import { clearAllSyncRetries } from "../../utils/syncRetry.js";
-
-/**
- * Moves every loaded query to `paused`, and to `error + paused` when the
- * connection failed permanently. Socket.IO keeps `active === true` while it is
- * still retrying, so that flag is what separates a temporary interruption from
- * a rejection the client will never recover from on its own.
- */
-export function pauseQueries<
-  D extends object,
-  S extends object,
-  O extends object,
->(
-  client: DocSyncClient<D, S, O>,
-  connectionError: DocSyncError | undefined,
-): void {
-  client["_connectionFetchStatus"] = "paused";
-  if (connectionError) client["_connectionError"] = connectionError;
-  for (const docId of client["_docsCache"].keys()) {
-    if (connectionError) {
-      dispatchDocQueryConnectionError(client, docId, connectionError);
-    } else {
-      dispatchDocQueryDisconnected(client, docId);
-    }
-  }
-}
 
 /** Tells the other tabs this client is gone from every document it had open. */
 function broadcastPresenceLeft<
