+++ tests/docsync/unit/client/queryResultReducer/queryResultReducer.test.ts
@@ -13,6 +13,16 @@ describe("createQueryResultReducer", () => {
   test.each(combinations)(
     "$stateCase.name + $actionCase.name",
     ({ stateCase, actionCase }) => {
+      // A settled query already holds the newest attempt's result, so a
+      // terminal network action from a superseded attempt leaves it untouched.
+      if (
+        actionCase.ignoredWhenSettled &&
+        stateCase.state.fetchStatus === "idle"
+      ) {
+        expect(actionCase.run(stateCase.state)).toBe(stateCase.state);
+        return;
+      }
+
       expect(actionCase.run(stateCase.state)).toStrictEqual(
         actionCase.expected(stateCase.state),
       );
